useSearchSchemaFields custom hook and refactored from useschemasettings
