+++ app/packages/state/src/hooks/index.ts
@@ -51,4 +51,5 @@ export { default as useUpdateSamples } from "./useUpdateSamples";
 export { default as useSchemaSettings } from "./useSchemaSettings";
 export { default as useSetShowNestedFields } from "./schema/useSetShowNestedFields";
 export { default as useSetSelectedFieldsStage } from "./schema/useSetSelectedFieldsStage";
+export { default as useSearchSchemaFields } from "./schema/useSearchSchemaFields";
 export { default as withSuspense } from "./withSuspense";
